Completed 6.6 and 6.7
